% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniCRAN.R
\name{add_CRAN_pkg}
\alias{add_CRAN_pkg}
\title{Add CRAN packages into miniCRAN}
\usage{
add_CRAN_pkg(package_name, miniCRAN_path, CRAN_url, package_types)
}
\arguments{
\item{package_name}{A package name to install}

\item{miniCRAN_path}{A path of miniCRAN}

\item{package_types}{Types of package to be installed}

\item{CRAN_path}{A url of CRAN mirror}
}
\description{
Add CRAN packages into miniCRAN
}
