% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEssential.R
\name{configure}
\alias{configure}
\title{Configure buildEssential}
\usage{
configure(miniCRAN_path, lib_path)
}
\arguments{
\item{miniCRAN_path}{A path of miniCRAN}

\item{lib_path}{A search path of the library tree which packages are stored}
}
\description{
BuildEssential sets two environmental variables. One is a miniCRAN path by
\code{options(repos)}, and another is a library path by \code{.libPaths}.
You'll need to put parameters in order to do this. If those paths are not
exist, then this function will create them.
}
