% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEssential.R
\name{get_packages}
\alias{get_packages}
\title{Get required package names}
\usage{
get_packages(description_path)
}
\arguments{
\item{description_paths}{A vector of DESCRIPTION paths}
}
\description{
Get required package names
}
