% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniCRAN.R
\name{init_miniCRAN}
\alias{init_miniCRAN}
\title{Initialize miniCRAN}
\usage{
init_miniCRAN(miniCRAN_path, CRAN_url, package_types)
}
\arguments{
\item{miniCRAN_path}{A path of miniCRAN}

\item{package_types}{Types of package to be installed}

\item{CRAN_path}{A url of CRAN mirror}
}
\description{
Initialize miniCRAN
}
