% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{install_package}
\alias{install_package}
\title{Install miniCRAN package}
\usage{
install_package(package_name)
}
\arguments{
\item{package_name}{A package name to install}
}
\description{
Install miniCRAN package
}
