% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{install_src_packages}
\alias{install_src_packages}
\title{Install source packages}
\usage{
install_src_packages(dir_path = NULL, pattern = "*.tar.gz")
}
\arguments{
\item{dir_path}{A search path of the directory which source packages are
stored}

\item{pattern}{A regex pattern for source packages. \code{*.tar.gz} by
default}
}
\description{
Install source packages
}
