% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildr.R
\name{setup}
\alias{setup}
\title{Setup Your Build Essentials}
\usage{
setup(description_path, minicran_path,
  cran_url = "http://cran.us.r-project.org", package_type = "source")
}
\arguments{
\item{description_path}{to DESCRIPTION}
}
\description{
Setup Your Build Essentials
}
