% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minicran.R
\name{update_cran_pkg}
\alias{update_cran_pkg}
\title{Update cran packages in minicran}
\usage{
update_cran_pkg(minicran_path, cran_url = "http://cran.us.r-project.org",
  package_types = "source")
}
\arguments{
\item{minicran_path}{A path of minicran}

\item{cran_url}{A url of cran mirror}

\item{package_types}{Types of package to be installed}
}
\description{
Update cran packages in minicran
}
