% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniCRAN.R
\name{update_CRAN_pkgs}
\alias{update_CRAN_pkgs}
\title{Update CRAN packages in miniCRAN}
\usage{
update_CRAN_pkgs(miniCRAN_path, CRAN_url, package_type)
}
\arguments{
\item{miniCRAN_path}{A path of miniCRAN}

\item{CRAN_url}{A url of CRAN mirror}

\item{package_types}{Types of package to be installed}
}
\description{
Update CRAN packages in miniCRAN
}
