% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minicran.R
\name{add_cran_pkg}
\alias{add_cran_pkg}
\title{Add cran packages into minicran}
\usage{
add_cran_pkg(package_name, minicran_path,
  cran_url = "http://cran.us.r-project.org", package_types = "source")
}
\arguments{
\item{package_name}{A package name to install}

\item{minicran_path}{A path of minicran}

\item{package_types}{Types of package to be installed}

\item{cran_path}{A url of cran mirror}
}
\description{
Add cran packages into minicran
}
