% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minicran.R
\name{init_minicran}
\alias{init_minicran}
\title{Initialize minicran}
\usage{
init_minicran(minicran_path, cran_url = "http://cran.us.r-project.org",
  package_types = "source")
}
\arguments{
\item{minicran_path}{A path of minicran}

\item{package_types}{Types of package to be installed}

\item{cran_path}{A url of cran mirror}
}
\description{
Initialize minicran
}
